Rewrite Message Configuration
Make use of the token system
Fix a few errors with variable parsing
Ensure we read the regex properly for variables like currency symbol

+++ src/main/java/com/nickimpact/GTS/Utils/UpdateLotsTask.java
@@ -10,12 +10,16 @@ import net.minecraft.server.MinecraftServer;
 import net.minecraft.util.text.TextFormatting;
 import org.spongepowered.api.Sponge;
 import org.spongepowered.api.entity.living.player.Player;
+import org.spongepowered.api.event.cause.Cause;
+import org.spongepowered.api.event.cause.NamedCause;
 import org.spongepowered.api.scheduler.Scheduler;
 import org.spongepowered.api.scheduler.Task;
+import org.spongepowered.api.service.economy.account.UniqueAccount;
 import org.spongepowered.api.text.Text;
 import org.spongepowered.api.text.format.TextColors;
 import org.spongepowered.api.text.format.TextStyles;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.Date;
 import java.util.HashMap;
@@ -56,7 +60,7 @@ public class UpdateLotsTask {
             HashMap<String, Optional<Object>> textOptions = Maps.newHashMap();
             textOptions.put("pokemon", Optional.of(lot.getItem().getName()));
 
-            player.get().sendMessage(MessageConfig.getMessage("GTS.Remove.Expired", textOptions));
+            player.get().sendMessage(MessageConfig.getMessage("Generic.Remove.Expired", textOptions));
             Optional<PlayerStorage> storage = PixelmonStorage.pokeBallManager.getPlayerStorageFromUUID((MinecraftServer)Sponge.getServer(), Sponge.getServer().getPlayer(item.getOwner()).get().getUniqueId());
             if(storage.isPresent()) {
                 storage.get().addToParty(item.getPokemon(lot, player.get()));
@@ -77,10 +81,30 @@ public class UpdateLotsTask {
                 textOptions.put("pokemon", Optional.of(item.getName()));
                 textOptions.put("price", Optional.of(lot.getStPrice()));
 
-                for(Text text : MessageConfig.getMessages("GTS.Auction.Award", textOptions))
-                    Sponge.getServer().getBroadcastChannel().send(text);
-                storage.get().addToParty(item.getPokemon(lot, player));
-                GTS.getInstance().getSql().deleteLot(lot.getLotID());
+                BigDecimal price = new BigDecimal(lot.getStPrice());
+                Optional<UniqueAccount> account = GTS.getInstance().getEconomy().getOrCreateAccount(player.getUniqueId());
+                if(account.isPresent()) {
+                    UniqueAccount acc = account.get();
+                    if (acc.getBalance(GTS.getInstance().getEconomy().getDefaultCurrency()).intValue() < price.intValue()) {
+                        for (Text text : MessageConfig.getMessages("Generic.Purchase.Error.Not Enough", null))
+                            player.sendMessage(text);
+                        return;
+                    }
+                    acc.withdraw(GTS.getInstance().getEconomy().getDefaultCurrency(), price, Cause.source(GTS.getInstance()).build());
+
+                    for (Text text : MessageConfig.getMessages("Auctions.Award", textOptions))
+                        Sponge.getServer().getBroadcastChannel().send(text);
+                    storage.get().addToParty(item.getPokemon(lot, player));
+                    GTS.getInstance().getSql().deleteLot(lot.getLotID());
+
+                    Optional<UniqueAccount> ownerAccount = GTS.getInstance().getEconomy().getOrCreateAccount(lot.getOwner());
+                    if(ownerAccount.isPresent()){
+                        UniqueAccount owner = ownerAccount.get();
+                        owner.deposit(GTS.getInstance().getEconomy().getDefaultCurrency(), price, Cause.of(NamedCause.source(GTS.getInstance())));
+                    } else {
+                        GTS.getInstance().getLogger().error("Player '" + Sponge.getServer().getPlayer(lot.getOwner()).get().getName() + "' was unable to receive $" + price.intValue() + " from the GTS");
+                    }
+                }
             } else {
                 GTS.getInstance().getLogger().error("An error occurred when trying to award a pokemon to " + player.getName());
             }
